Clean up of search engine classes
- Separate OpenSearchEngine into it's own file
- Separate OpenSearchParser into it's own file
- Move file under a SearchEngines folder, do the same for tests files
- Order function with MARK (private under private, etc)
- Remove capital letters from constants
- Make things enclosed in classes when possible instead of global
- Use standard comment format with ///
- Remove class func to normal func for getUnorderedBundledEnginesFor and possibilitiesForLanguageIdentifier as they are used under a normal func
